Feat/quiz UI

+++ app/src/main/java/com/beside/hackathon/data/model/quiz/Option.kt
@@ -1,6 +1,6 @@
 package com.beside.hackathon.data.model.quiz
 
 data class Option(
-    val optionId: Long,
+    val optionId: Int,
     val content: String,
 )
